frameworks/motia/motia-py/packages/motia/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "motia"
version = "1.0.4rc1"
description = "Motia framework for III Engine"
authors = [{ name = "III" }]
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]



dependencies = [
    "iii-sdk",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tomli>=2.0; python_version < \"3.11\"",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.8",
    "ruff>=0.2",
    "httpx>=0.27",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
]

[project.scripts]
motia = "motia.cli:main"
motia-iii = "motia.cli:main"

[tool.uv.sources]
iii-sdk = { path = "../../../../../sdk/packages/python/iii" }

[tool.hatch.build.targets.wheel]
packages = ["src/motia"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["iii", "iii.*", "watchfiles", "watchfiles.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestDeprecationWarning",
]
markers = [
    "integration: marks tests as integration tests (require III engine)",
]

[tool.coverage.run]
branch = true
source = ["src/motia"]

[tool.coverage.report]
show_missing = true
skip_covered = true
